feat/#2 Functional color display
The color display that shows the color one inputs now works. Displays color if input is valid, else shows a standard color.

The function that writes to css now uses the color that is displayed next to the input fields, whether it's the user input or standard color. This makes sure one does not need to check if input is valid when writing the css as it is already checked when displaying the color.